# specs/004-submission-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Query Dungeon Submission API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
  /api/challenges:
    get:
      summary: List participant-safe challenges
      responses:
        "200":
          description: Challenge list
          content:
            application/json:
              schema:
                type: object
                required: [challenges]
                properties:
                  challenges:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChallengeSummary"
  /api/submissions:
    post:
      summary: Create and evaluate a submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
      responses:
        "201":
          description: Well-formed submission accepted, stored, and evaluated; SQL safety, size, and timeout failures are returned here with failed status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"
        "400":
          description: Malformed JSON or request-schema failure before a submission is accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown challenge, variant, or submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/submissions/{id}:
    get:
      summary: Fetch submission status and result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Submission status and result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"
        "404":
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/challenges/{challengeId}/leaderboard:
    get:
      summary: List best correct completed submissions for a challenge
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
        - name: variant
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: object
                required: [challengeId, variant, entries]
                properties:
                  challengeId:
                    type: string
                  variant:
                    type: [string, "null"]
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/LeaderboardEntry"
        "404":
          description: Challenge or variant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    CreateSubmissionRequest:
      type: object
      required: [challengeId, sql]
      properties:
        challengeId:
          type: string
        variant:
          type: string
        participantName:
          type: string
          minLength: 1
        participantId:
          type: string
          minLength: 1
        sql:
          type: string
          minLength: 1
          maxLength: 65536
        notes:
          type: string
      anyOf:
        - required: [participantName]
        - required: [participantId]
      additionalProperties: false
    SubmissionStatus:
      type: string
      enum: [pending, running, completed, failed]
    SubmissionResult:
      type: object
      required:
        - submissionId
        - status
        - correctness
        - latencyMs
        - executionTimeMs
        - planningTimeMs
        - rowsReturned
      properties:
        submissionId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        correctness:
          type: [boolean, "null"]
        latencyMs:
          type: [number, "null"]
        executionTimeMs:
          type: [number, "null"]
        planningTimeMs:
          type: [number, "null"]
        rowsReturned:
          type: [integer, "null"]
        errorMessage:
          type: [string, "null"]
    ChallengeSummary:
      type: object
      required: [id, title, difficulty, variants]
      properties:
        id:
          type: string
        title:
          type: string
        difficulty:
          type: string
        variants:
          type: array
          items:
            type: object
            required: [id, title, difficulty]
            properties:
              id:
                type: string
              title:
                type: string
              difficulty:
                type: string
    LeaderboardEntry:
      type: object
      required:
        - submissionId
        - participantName
        - participantId
        - latencyMs
        - executionTimeMs
        - planningTimeMs
        - rowsReturned
        - submittedAt
      properties:
        submissionId:
          type: string
          format: uuid
        participantName:
          type: [string, "null"]
        participantId:
          type: [string, "null"]
        latencyMs:
          type: [number, "null"]
        executionTimeMs:
          type: [number, "null"]
        planningTimeMs:
          type: [number, "null"]
        rowsReturned:
          type: integer
        submittedAt:
          type: string
          format: date-time
        completedAt:
          type: [string, "null"]
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string